refactor toolbar
