+++ packages/web/src/index.ts
@@ -23,13 +23,14 @@ export * from './formInputs/toggle'
 export * from './formInputs/upload/'
 export * from './grid/'
 export * from './label'
-export * from './layout/accordion/'
-export * from './layout/card'
-export * from './layout/container'
-export * from './layout/flex'
-export * from './layout/horizontal'
-export * from './layout/panel'
-export * from './layout/vertical'
+export * from './layouts/accordion/'
+export * from './layouts/card'
+export * from './layouts/container'
+export * from './layouts/flex'
+export * from './layouts/horizontal'
+export * from './layouts/panel'
+export * from './layouts/toolbar'
+export * from './layouts/vertical'
 export * from './mixins/'
 export * from './notification'
 export * from './richText'
