+++ packages/web/src/layouts/accordion/accordionPanel.ts
