+++ packages/web/src/layouts/horizontal.ts
