+++ packages/web/src/layouts/vertical.ts
